% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_IBD.R
\name{calc_IBD}
\alias{calc_IBD}
\alias{calc_IBD.numeric}
\alias{calc_IBD.matrix}
\title{Calculation of IBD}
\usage{
calc_IBD(dos, p1hom, p2hom, ploidy = 2)

\method{calc_IBD}{numeric}(dos, p1hom, p2hom, ploidy = 2)

\method{calc_IBD}{matrix}(dos, p1hom, p2hom, ploidy = 2)
}
\arguments{
\item{dos}{integer or matrix containing genotypes (dosages of allele A, from 0 to ploidy).
If `dos` is a matrix, each row must be labelled with a marker name, and each column is
interpreted as an individual.}

\item{p1hom}{numeric vector or matrix indicating which parental homologue, from parent 1, contains allele A.
If `dos` is an integer, `p1hom` must be a vector of length ploidy. If `dos` is a matrix,
`p1hom` must be a matrix of ncol = ploidy and nrow = nrow(dos) and the same rownames
as `dos`.}

\item{p2hom}{numeric vector or matrix indicating which parental homologue, from parent 2, contains allele A.
If `dos` is an integer, `p2hom` must be a vector of length ploidy. If `dos` is a matrix,
`p2hom` must be a matrix of ncol = ploidy and nrow = nrow(dos) and the same rownames
as `dos`.}

\item{ploidy}{numeric, indicating the ploidy of an organism. Only even ploidies allowed and
all individuals are expected to have the same ploidy.}
}
\value{
The na攼㹦ve probability of having inherited each parental homologue. If the
offspring dosage is impossible given the parental homologues, NA is returned.\n
If `dos`is a vector, returns a one-row matrix, each column containing the
probability of having inherited that parental homologue (the first columns corresponding
to parent 1, and the rest to parent 2). \n If `dos` is a matrix, a list of matrices is returned.
Each element is the probabilities for all individuals of having inherited one parental
homologue. There should be ploidy*2 elements in the list.
}
\description{
Calculation of IBD

Calculation of IBD for a single data point

Calculation of IBD for multiple markers and/or individuals
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Function for a single marker of a single individual.

\item \code{matrix}: Function for handling matrix input to calc_IBD
}}

