% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_IBD.R
\name{config}
\alias{config}
\title{Configuration function}
\usage{
config(g, ploidy = 2)
}
\arguments{
\item{g}{integer, a genotype <= ploidy and >=0}

\item{ploidy}{integer, indicating ploidy.}
}
\value{
a list of vectors, where each element indicates
one possible parental chromosome inheritance configuration.
}
\description{
Small internal function that returns all possible configurations of
parental inheritance given a genotype
}
