% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{custom_calc.maps.sphere}
\alias{custom_calc.maps.sphere}
\title{3-dimensional MDS-mapping}
\usage{
custom_calc.maps.sphere(
  fname,
  p = 100,
  n = NULL,
  weightfn = "lod2",
  mapfn = "haldane"
)
}
\arguments{
\item{fname}{Character string specifying the base name of the file fname.txt
which contains the data to be analysed.
The file should be white space or tab separated.}

\item{p}{Integer - the penalty for deviations from the sphere - higher p
forces points more closely onto a sphere.}

\item{n}{Vector of integers or strings containing markers to be omitted from the analysis.}

\item{weightfn}{Character string specifying the values to use for
the weight matrix in the MDS 'lod2' or 'lod'.}

\item{mapfn}{Character string specifying the map function to use on the recombination
fractions 'haldane' is default, 'kosambi' or 'none'.}
}
\value{
Estimated linkage map. See `MDSMap::calc.maps.sphere()` for further details.
}
\description{
Custom function identical to `MDSMap::calc.maps.sphere()` except for a reduction
in the number of iterations used from 1e06 to 2e04, which greately reduces computational
time with little change in the final outcome.
}
\keyword{internal}
