% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documents.R
\docType{data}
\name{geno}
\alias{geno}
\title{Genotype matrix}
\usage{
data("genotype")
}
\description{
A genotype matrix is a markers x individuals matrix, with marker names as row names
and individual names as column names. Dosages (as integers) are expected.
This example is from one chromosome of a diploid organism (strabwerry).
}
\author{
Alejandro Therese Navarro \email{alejandro.theresenavarro@wur.nl}
}
\keyword{data}
