% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{graphical_genotype}
\alias{graphical_genotype}
\alias{graphical_genotype.matrix}
\alias{graphical_genotype.list}
\title{Graphical genotype plotter}
\usage{
graphical_genotype(
  IBD,
  only_informative = F,
  cex.axis = 0.8,
  main = NULL,
  non_inf = c(0.3, 0.7),
  ...
)

\method{graphical_genotype}{matrix}(
  IBD,
  only_informative = F,
  cex.axis = 0.8,
  main = NULL,
  non_inf = c(0.3, 0.7),
  ...
)

\method{graphical_genotype}{list}(
  IBD,
  only_informative = F,
  cex.axis = 0.8,
  main = NULL,
  non_inf = c(0.3, 0.7),
  ...
)
}
\arguments{
\item{IBD}{IBD matrix or list of IBD matrices. If IBD is a list, the
names are used as subplot titles.}

\item{only_informative}{logical, should only informative markers be plotted?
False by default.}

\item{cex.axis}{size of the bottom labels.}

\item{main}{title to display in the plot.}

\item{non_inf}{numeric, lower and upper probability boundaries to consider an
IBD probability non-informative (if they fall within the threshold they will be
ignored during prediction). Defaults to 0.3 - 0.7. Symmetrical boundaries are
recommended but not necessary.}

\item{...}{Additional parameters to be passed to `plot`. For instance
`cex.axis` or `main`.}
}
\value{
A plot containing graphical representations of IBD. Dark
is a probability close to 1 and light is a probability close to 0.
}
\description{
Given an IBD matrix or a list of matrices, a graphical genotype
plot is returned. Graphical genotypes allow to assess which homologue
an individual has obtained, they highlight putative errors and show
non-informative genotypes.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: function for a single matrix.

\item \code{list}: function for a list of matrices.
}}

\examples{
data("smooth_descent")
IBD <- sdescent$predIBD

#One single homologue
graphical_genotype(IBD[[1]], only_informative = TRUE)
#Or all homologues
graphical_genotype(IBD, only_informative = TRUE)
}
