% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documents.R
\docType{data}
\name{hom}
\alias{hom}
\title{Homologue matrix}
\usage{
data("homologue")
}
\description{
A homologue matrix is a markers x homologues matrix, with marker names
as row names and homologue names (as many as ploidy for each parent) for
columns. This example is from one chromosome of a diploid organism (strabwerry).
}
\author{
Alejandro Therese Navarro \email{alejandro.theresenavarro@wur.nl}
}
\keyword{data}
