% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documents.R
\docType{data}
\name{linkdf}
\alias{linkdf}
\title{Linkage data.frame}
\usage{
data("linkage")
}
\description{
This linkage data.frame contains pairwise linkage information as obtained by
`linkdf_shortcut()`, which calculates linkage using polymapR. It contains coluns
marker_a, marker_b, and LOD_independence, which is used (rather than LOD score).
}
\author{
Alejandro Therese Navarro \email{alejandro.theresenavarro@wur.nl}
}
\keyword{data}
