% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{linkdf_shortcut}
\alias{linkdf_shortcut}
\title{Linkage estimation}
\usage{
linkdf_shortcut(geno, p1name, p2name, ploidy, ncores = 1)
}
\arguments{
\item{geno}{Genotype matrix with individuals on columns and markers on rows. Colum names
are expected.}

\item{p1name}{Column name of the first parent.}

\item{p2name}{Column name of the second parent.}

\item{ploidy}{Ploidy of the population (only even ploidies identical for the whole population
are allowed)}

\item{ncores}{Number of cores to use in the linkage calculation. It is recommended to
set this number at most to the number of available cores - 1.}
}
\value{
Data.frame containing recombination frequency, LOD estimate and LOD for independence
for each pair of markers. See `polymapR::linkage()` for more information.
}
\description{
The `polymapR::linkage()` allows to calculate pair-wise linkage estimates between markers
using the `MDSmap` package. However, it requires to be run once for each marker pair
combination (for 1x0 - 1x0, 1x0 - 1x1, 1x1 - 0x1 ...). To simplify this process one can
use this shortcut, which analyzes the present genotypes and launches `linkage()` as many
times as necessary to obtain all meaningful pair-wise estimates (for instance, 1x0 - 0x1
pairs are ignored since they are always independent). It returns a data-frame
with all pair-wise links.
}
\keyword{internal}
