% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{linkpar}
\alias{linkpar}
\title{Linkage parameter calculator}
\usage{
linkpar(genop1, genop2, pnames)
}
\arguments{
\item{genop1}{Numeric vector containing genotypes of parent 1}

\item{genop2}{Numeric vector containing genotypes of parent 2}

\item{pnames}{Character vector of length 2 containing parent names}
}
\value{
A list where each element contains the parameters `markertype1`,
`markertype2`, `target_parent` and `other_parent` for each type of marker
defined by the combination of P1 and P2 genotypes.
}
\description{
This function aids in the definition of parameters to pass to the
`linkage` function in the package `polymapR`.
}
\keyword{internal}
