% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documents.R
\docType{data}
\name{map}
\alias{map}
\title{Map data.frame}
\usage{
data("map")
}
\description{
A map data.frame should contain at least two columns: marker (with marker names)
and position, with physical or genetic positions. Each row corresponds to a marker.
This example is from one chromosome of a diploid organism (strabwerry).
}
\author{
Alejandro Therese Navarro \email{alejandro.theresenavarro@wur.nl}
}
\keyword{data}
