% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mdsmap}
\alias{mdsmap}
\title{MDSmap caller}
\usage{
mdsmap(linkdf, ndim = 2)
}
\arguments{
\item{linkdf}{Linkage data.frame as defined in the `MDSMap::estimate.map()` function.}

\item{ndim}{2 or 3, number of dimensions to use in MDS-mapping.}
}
\value{
Estimated linkage map, including position and other parameters.
}
\description{
Shortcut function to launch MDSmap
}
\keyword{internal}
