% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_IBD.R
\name{naive_IBD.numeric}
\alias{naive_IBD.numeric}
\alias{naive_IBD.matrix}
\title{Calculation of IBD for a single data point}
\usage{
\method{naive_IBD}{numeric}(geno, p1hom, p2hom, ploidy = 2, ...)

\method{naive_IBD}{matrix}(geno, p1hom, p2hom, ploidy = 2, p1name = "P1", p2name = "P2", ...)
}
\description{
Calculation of IBD for a single data point

Calculation of IBD for multiple markers and/or individuals
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Function for a single marker of a single individual.

\item \code{matrix}: Function for handling matrix input to naive_IBD
}}

