% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_IBD.R
\name{predict_IBD}
\alias{predict_IBD}
\alias{predict_IBD.list}
\alias{predict_IBD.matrix}
\alias{predict_IBD.numeric}
\title{Predict IBD}
\usage{
predict_IBD(IBD, map, interval = 10)

\method{predict_IBD}{list}(IBD, map, interval = 10)

\method{predict_IBD}{matrix}(IBD, map, interval = 10)

\method{predict_IBD}{numeric}(IBD, map, interval = 10)
}
\arguments{
\item{IBD}{observed IBD. Can be given as a vector containing the
observed IBDs of one individual, a single matrix (markers x individual) or a list
of matrices. Matrices must contain rownames and columnames.}

\item{map}{data.frame containing "marker" and "position" columns at least.}

\item{interval}{numeric indicating the interval size to be used in prediction.}
}
\value{
the output will have the same format as the input given. That is, an input
IBD vector will return an IBD vector, a matrix will return a matrix and a list of matrices
will return a list of matrices.
}
\description{
IBD prediction algorithm, performs a distance-informed weighted average using
a genetic map. IBD probabilities between 0.3 and 0.7 are ignored as non-informative.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Function for a list of matrices

\item \code{matrix}: Function for a single matrix

\item \code{numeric}: Function for a numeric vector
}}

