% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{rec_count}
\alias{rec_count}
\alias{rec_count.matrix}
\alias{rec_count.list}
\title{Recombination counter}
\usage{
rec_count(ibd, map, non_inf = c(0.3, 0.7))

\method{rec_count}{matrix}(ibd, map, non_inf = c(0.3, 0.7))

\method{rec_count}{list}(ibd, map, non_inf = c(0.3, 0.7))
}
\arguments{
\item{ibd}{IBD matrix or list of matrices. Must contain row.names indicating
the marker name.If `ibd` is a list, it must contain matrices for all parental homologues.}

\item{map}{data.frame with at least columns "position" and "marker". It is advisable
to use a single}

\item{non_inf}{Interval determining what is to be considered a non informative
probability. Values below non_inf[1] will be treated as 0 and above non_inf[2]
will be treated as 1. Values between the two will be ignored.}
}
\value{
A list containing:
* individual: recombination counts per individual. Provided per parent if
`ibd` was a list.
* on_map: a data.frame with recombination counts across individuals per
window. Provided per parent if `ibd` was a list.
}
\description{
Using IBD matrices, one can estimate the number of recombinations in a single chromosome. This
function calculates the number of recombinations using a single IBD matrix
or a list of IBD matrices of all parental homologues. Recombinations are
calculated per individual and per window. A genetic map is required.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Recombination count for a single IBD matrix.

\item \code{list}: Recombination count for a list of IBD matrices
for all parental homologues.
}}

