% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{recdist_plot}
\alias{recdist_plot}
\title{Recombination-distance plotter}
\usage{
recdist_plot(recdist, ...)
}
\arguments{
\item{recdist}{data.frame containing at least columns "recombination" and "distance".
Can be obtained from `recdist_calc`.}

\item{...}{additional parameters passed onto `plot`}
}
\description{
Plots a recombination-distance plot that allows to assess whether a given map order
matches well with recombination estimates,
since larger recombination estimates should be equivalent to
larger map distances. It is useful to colour the dots with
the LOD score.
}
