% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{reorder_tau}
\alias{reorder_tau}
\title{Reorder parameter calculator}
\usage{
reorder_tau(oldmap, newmap)
}
\arguments{
\item{oldmap}{a data.frame containing at least "marker" and "position" columns.}

\item{newmap}{a data.frame containing at least "marker" and "position" columns.
All marker names must be present in `oldmap`.}
}
\value{
A single parameter indicating the rank order correlation.
}
\description{
Given two maps with the same markers, calculates the
reordering parameter \eqn{\tau} between them. This is based on the
rank-correlation, or Kendall correlation.
}
\examples{
data("iter_smooth_descent")
old <- sd_iter$iter1$oldmap
new <- sd_iter$iter1$newmap

reorder_tau(old,new)

}
