% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{rowfuse}
\alias{rowfuse}
\title{Matrix merging of rows with equal names}
\usage{
rowfuse(mat1, mat2, fill = NA)
}
\arguments{
\item{mat1}{matrix 1, with dimnames.}

\item{mat2}{matrix 2, with dimnames.}

\item{fill}{value to put inside the output matrix for those rows
that were present in only one of the two matrices.}
}
\value{

}
\description{
Similar to merge, given two matrices, row names are matched
and columns are bound. All columns are preserved and all row.names
present in both matrices are present in the result. Empty cells,
belonging to a row that is present in one matrix and not the other
are filled in with the `fill` value-
}
