% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documents.R
\docType{data}
\name{sd_iter}
\alias{sd_iter}
\title{Smooth Descent iterations output}
\usage{
data("iter_smooth_descent")
}
\description{
Example output from 5 iterations of Smooth Descent (using `smooth_map_iter()`).
The example data from this package were used (`geno`,`hom`,`map`). Mapping was
performed with `ndim = 3`. The rest of parameters were kept on default.
}
\author{
Alejandro Therese Navarro \email{alejandro.theresenavarro@wur.nl}
}
\keyword{data}
