% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documents.R
\docType{data}
\name{sdescent}
\alias{sdescent}
\title{Smooth Descent output}
\usage{
data("smooth_descent")
}
\description{
Example output from a single run of the Smooth Descent algorithm (using `smooth_descent()`).
The example data from this package were used (`geno`,`hom`,`map`). The rest of
parameters were kept on default.
}
\author{
Alejandro Therese Navarro \email{alejandro.theresenavarro@wur.nl}
}
\keyword{data}
