% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_wrapper.R
\name{smooth_map}
\alias{smooth_map}
\title{Smooth Mapping}
\usage{
smooth_map(
  geno,
  homologue,
  map = NULL,
  ploidy = 2,
  p1name = NULL,
  p2name = NULL,
  prediction_interval = 10,
  error_threshold = 0.8,
  ncores = 10,
  mapping_ndim = 2,
  estimate_premap = F,
  verbose = T
)
}
\arguments{
\item{geno}{matrix with markers on the rows, individuals on the columns.
Rownames and columnames expected.}

\item{homologue}{matrix with markers on the rows, homologue names on the columns.
Rownames and columnames expected.}

\item{map}{optionally, data.frame with at least columns "marker" and "position". If it is
not specified, a map will be estimated from the uncorrected genotype data with polymapR.}

\item{ploidy}{numeric indicating the ploidy. Both parents must be of the same ploidy,
and it is assumed that "homologue" has 2*ploidy columns.}

\item{p1name}{character, name of the first parent. Must be present in the geno columnames.
If it's not specified it will be taken as the name of the first column.}

\item{p2name}{character, name of the second parent. Must be present in the geno columnames.
If it's not specified it will be taken as the name of the second column.}

\item{prediction_interval}{numeric, interval to be used during the IBD prediction
step. It should be specified in the same units as the "position" in the map.}

\item{error_threshold}{numeric, threshold over which a marker is considered erroneous.
Usually 0.8 should be good enough to be sensitive but stringent (not have false positives)}

\item{ncores}{number of cores to use for linkage estimation.}

\item{mapping_ndim}{2 or 3, number of dimensions to use for multi-dimensional mapping}

\item{estimate_premap}{logical, whether to use polymapR to estimate a preliminary map.}

\item{verbose}{logical, should smooth descent report the steps it takes?}
}
\value{
list containing the following items:
* obsIBD: list of observed IBD matrices (marker x individual) for each parental homologue
* predIBD: list of predicted IBD matrices (marker x individual) for each parental homologue
* oldmap: data.frame containing the original map
* error: list of error matrices (marker x individual) for each parental homologue
* newIBD: list of observed IBD matrices of the corrected genotypes.
* newmap: data.frame containing the new updated map.
* newgeno: matrix containing the new genotypes.
* rec: list containing the recombination counts using the observed IBD (`obs`),
the predicted IBD (`pred`) and the updated IBD (`new`). `obs` and `pred` use the old
map order while `new` uses the re-estimated map. For more information see `rec_count()`.
* recdist: data.frame containing pair-wise recombination and distance between markers,
useful to plot using `recdist_plot()`
* tau: reordering parameter tau (Kendall's rank correlation). Obtained with `reorder_tau()`
* eliminated: markers eliminated due to a large average neighbour distance, they tend
to be problematic when re-mapping.
}
\description{
This function allows to apply one single interation of the Smooth Descent
algorithm, including a re-mapping step using new genotypes. Optionally, the
preliminary map can also be estimated instead of being provided. Both mapping
procedures are performed with packages `polymapR` and `MDSMap`, which perform
multi-dimensional scaling mapping. For usage of the Smooth Descent algorithm
with a different map algorithm see `smooth_descent()`, which performs the
genotype correction without re-mapping.
}
