% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{extract}
\alias{extract}
\title{Extraction from list}
\usage{
extract(l, index = 1, simplify = T)
}
\arguments{
\item{l}{list of elements with identical structure}

\item{index}{numeric, character or logical to be applied to each element of list}

\item{simplify}{should the output be simplified}
}
\value{

}
\description{
Given a list of identical objects, extract the same index of each object.
This is a shortcut for `sapply(list,'[[',index)`
}
