% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{extract}
\alias{extract}
\title{Extraction from list}
\usage{
extract(l, index = 1, simplify = T)
}
\arguments{
\item{l}{list of elements with identical structure}

\item{index}{numeric, character or logical to be applied to each element of list}

\item{simplify}{should the output be simplified}
}
\value{
list, vector or matrix of extracted objects
}
\description{
Given a list of identical objects, extract the same index of each object.
This is a shortcut for `sapply(list,'[[',index)`
}
\examples{
example <- list(data.frame(A = "first dataframe", B = "second argument"),
data.frame(A = "second dataframe", B = "second argument"))
extract(example,"A")
}
