% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{graphical_genotype}
\alias{graphical_genotype}
\alias{graphical_genotype.matrix}
\alias{graphical_genotype.list}
\title{Graphical genotype plotter}
\usage{
graphical_genotype(IBD, only_informative = F, ...)

\method{graphical_genotype}{matrix}(IBD, cex.axis = 0.8, ...)

\method{graphical_genotype}{list}(IBD, cex.axis = 0.8, main = NULL, ...)
}
\arguments{
\item{IBD}{IBD matrix or list of IBD matrices. If IBD is a list, the
names are used as subplot titles.}

\item{only_informative}{logical, should only informative markers be plotted?
False by default.}

\item{...}{Additional parameters to be passed to `plot`. For instance
`cex.axis` or `main`.}
}
\value{
A plot containing graphical representations of IBD. Dark
is a probability close to 1 and light is a probability close to 0.
}
\description{
Given an IBD matrix or a list of matrices, a graphical genotype
plot is returned. Graphical genotypes allow to assess which homologue
an individual has obtained, they highlight putative errors and show
non-informative genotypes.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: function for a single matrix.

\item \code{list}: function for a list of matrices.
}}

