% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{iterplot}
\alias{iterplot}
\title{Plotting of map comparisons}
\usage{
iterplot(maplist, ...)
}
\arguments{
\item{maplist}{A list of maps containing at least "position" and "marker" columns.
The map from a single chromosome is expected. Names of this list are used as
axis names.}

\item{...}{Additional parameters to be passed to `plot()` such as `main`
for the title.}
}
\value{
A comparison plot between maps.
}
\description{
Given a list of maps with the same markers (re-orderings of the same markers)
it plots all maps with segments connecting the same marker as it moves through the
maps. Additionally, the reorder parameter tau (see `reorder_tau`) is also
printed between map comparisons.
}
\examples{
maplist <- list(pre = read.table("test/test_map.txt"),
flat = readRDS("test/test_flat_map.RDS")$locimap,
sphere = readRDS("test/test_sphere_map.RDS")$locimap)
maplist$flat$marker <- maplist$flat$locus
maplist$sphere$marker <- maplist$sphere$locus

iterplot(maplist,main = "Example")

}
