% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{linkdf_shortcut}
\alias{linkdf_shortcut}
\title{Linkdf shortcut}
\usage{
linkdf_shortcut(geno, ploidy, p1name, p2name, ncores = 1)
}
\arguments{
\item{geno}{numeric data.frame or matrix, with marker names as rownames and
individual names as column names.}

\item{ploidy}{numeric, ploidy of both parents.}

\item{p1name}{character, name of the first parent.}

\item{p2name}{characther, name of the second parent.}

\item{ncores}{numeric, number of threads to use in the linakge calculation.}
}
\value{
a linkage dataframe. Some estimates might be duplicated,
since they are calculated for both parents.
}
\description{
Shortcut to calculate the linkage data.frame using polymapR's approach
to autopolyploid linkage calculation. Random pairing is assumed, it only
considers one linkage group at a time and returns linkage both from
parent1 to parent2 and viceversa. Importantly, markers for which
parental genotype does not match offspring segregation are eliminated.
}
